+++ R/pipeline-docs.R
@@ -64,6 +64,13 @@
 #' how often to check the pipeline
 #' @param progress_title,progress_max,progress_quiet control the progress,
 #' see \code{\link[dipsaus]{progress2}}.
+#' @param destroy what part of data repository needs to be cleaned
+#' @param ask whether to ask
+#' @param format format of the extended data, default is \code{'json'}, other
+#' choices are \code{'yaml'}, \code{'fst'}, \code{'csv'}, \code{'rds'}
+#' @param error_if_missing,default_if_missing what to do if the extended data
+#' is not found
+#' @param data extended data to be saved
 #' @param ... other parameters, targets, etc.
 #' @return \describe{
 #' \item{\code{pipeline_root}}{the root directories of the pipelines}